configure.in: Bump to version 2.0.17.
libsrc/ParseArgv.c (ParseArgv): Change ARGV_CONSTANT to treat
src and dst and pointer to integer, as documented.  Reverts change
6.9 of this file.

+++ libsrc/ParseArgv.c
@@ -174,7 +174,7 @@ ParseArgv(argcPtr, argv, argTable, flags)
       infoPtr = matchPtr;
       switch (infoPtr->type) {
       case ARGV_CONSTANT:
-         *((long *) infoPtr->dst) = (long) infoPtr->src;
+         *((int *) infoPtr->dst) = (int) infoPtr->src;
          break;
       case ARGV_INT:
          nargs = (int) infoPtr->src;
